--- src/pages/Home.jsx
@@ -1,9 +1,0 @@
-function Home() {
-  return (
-    <div className="h-80 container mx-auto px-8 text-3xl text-center flex items-center justify-center">
-      Home
-    </div>
-  );
-}
-
-export default Home;
